feat: use meta tag for csp
