+++ src/pages/_document.jsx
@@ -1,35 +1,5 @@
 import { GTM_ID } from "@/src/config/constants";
 import Document, { Html, Head, Main, NextScript } from "next/document";
-import http from "http";
-
-const connectSources = [
-  process.env.NEXT_PUBLIC_MUMBAI_SUBGRAPH_URL,
-  process.env.NEXT_PUBLIC_FUJI_SUBGRAPH_URL,
-  process.env.NEXT_PUBLIC_API_URL,
-  "https://api.thegraph.com/ipfs/",
-  "https://ipfs.infura.io:5001/",
-]
-  .map((x) => (x || "").trim())
-  .filter((x) => !!x)
-  .join(" ");
-
-const csp = [
-  `script-src 'self' https://tagmanager.google.com https://*.googletagmanager.com`,
-  `connect-src 'self' https://*.neptunemutual.com/ https://*.google-analytics.com https://*.analytics.google.com https://*.googletagmanager.com ${
-    connectSources || ""
-  }`,
-  "style-src 'self' 'unsafe-inline' https://tagmanager.google.com https://fonts.googleapis.com",
-  "upgrade-insecure-requests",
-  "frame-ancestors 'none'",
-  "default-src 'none'",
-  "prefetch-src 'self'",
-  "manifest-src 'self'",
-  "base-uri 'none'",
-  "form-action 'none'",
-  "object-src 'none'",
-  "img-src 'self' www.googletagmanager.com https://ssl.gstatic.com https://www.gstatic.com https://*.google-analytics.com https://*.googletagmanager.com data:",
-  "font-src 'self' https://fonts.gstatic.com data:",
-];
 
 class MyDocument extends Document {
   static async getInitialProps(ctx) {
@@ -53,16 +23,9 @@ class MyDocument extends Document {
   render() {
     const { pageProps } = this.props;
 
-    console.log({ pageProps, http });
     return (
       <Html>
         <Head nonce={pageProps?.nonce}>
-          <meta
-            httpEquiv="Content-Security-Policy"
-            content={csp
-              .join("; ")
-              .replace(`script-src`, `script-src 'nonce-${pageProps?.nonce}'`)}
-          />
           <link
             rel="apple-touch-icon"
             sizes="57x57"
